+++ client/server/userController.ts
@@ -1,22 +1,6 @@
 import * as express from 'express';
-import { connect, ConnectOptions } from 'mongoose';
-import userModels from './userModels';
-
-const MONGO_URI =
-  'mongodb+srv://dbUser:codesmith@cluster0.drsef.mongodb.net/saamsa?retryWrites=true&w=majority';
-
-const SALT_WORK_FACTOR = 10;
 const bcrypt = require('bcryptjs');
-
-connect(MONGO_URI, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-  dbName: 'saamsa',
-} as ConnectOptions)
-  .then(() => console.log('Connected to MongoDB'))
-  .catch((err: Error) =>
-    console.log(`Error found inside the mongoose connect method: ${err}`)
-  );
+import userModels from './userModels';
 
 type userController = {
   createUser: (
