+++ package.json
@@ -35,6 +35,7 @@
     "@electron-forge/maker-squirrel": "^6.0.0-beta.61",
     "@electron-forge/maker-zip": "^6.0.0-beta.61",
     "@types/d3": "^7.0.0",
+    "@types/d3-scale": "^4.0.1",
     "@types/electron": "^1.6.10",
     "@types/enzyme": "^3.10.9",
     "@types/enzyme-adapter-react-16": "^1.0.6",
@@ -42,26 +43,26 @@
     "@types/jest": "^27.0.2",
     "@types/kafkajs": "^1.9.0",
     "@types/mongoose": "^5.11.97",
-    "@types/d3-scale": "^4.0.1",
     "@types/react": "^17.0.27",
     "@types/react-dom": "^17.0.9",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.3",
     "@typescript-eslint/eslint-plugin": "^4.32.0",
     "@typescript-eslint/parser": "^4.32.0",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.3",
     "babel-jest": "^27.2.4",
     "electron": "^15.1.1",
     "eslint": "^7.32.0",
     "eslint-plugin-react": "^7.26.1",
     "kafkajs": "^1.15.0",
-    "ts-jest": "^27.0.5",
-    "webpack-cli": "^4.8.0",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "ts-jest": "^27.0.5",
+    "webpack-cli": "^4.8.0"
   },
   "dependencies": {
     "@babel/core": "^7.15.5",
     "@babel/preset-env": "^7.15.6",
     "@babel/preset-react": "^7.14.5",
+    "@types/supertest": "^2.0.11",
     "axios": "^0.22.0",
     "babel-loader": "^8.2.2",
     "bcryptjs": "^2.4.3",
@@ -76,6 +77,7 @@
     "sass": "^1.42.1",
     "sass-loader": "^12.1.0",
     "style-loader": "^3.3.0",
+    "supertest": "^6.1.6",
     "typescript": "^4.4.3",
     "webpack": "^5.56.0",
     "webpack-dev-server": "^4.3.0"
